notebook/requirements.txt
amqp==5.1.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.6.0
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
billiard==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
celery==5.3.4
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-rest-auth==4.0.0
Django==4.1.7
django-allauth==0.52.0
django-bootstrap5==23.1
django-cors-headers==4.2.0
django-csv==1.2
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-excel==2.4.0
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
feedparser==6.0.8
filelock==3.9.0
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mutagen==1.47.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.1.1
ply==3.11
prompt-toolkit==3.0.39
protobuf==3.12.4
psutil==5.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycryptodomex==3.19.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python3-openid==3.2.0
pythran==0.10.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
scipy==1.8.0
SecretStorage==3.3.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
sympy==1.9
systemd-python==234
tartube==2.4.412
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
vine==5.0.0
virtualenv==20.21.0
wadllib==1.3.6
wcwidth==0.2.8
webencodings==0.5.1
websockets==11.0.3
whitenoise==6.5.0
xdg==5
xkit==0.0.0
xlwt==1.3.0
youtube-dl==2021.12.17
yt-dlp==2023.9.24
zipp==1.0.0
